// src/app/(protected)/(common)/member/claims/models/api/claimsResponse.ts
export type ClaimsResponse = {
  claims: Claim[];
};

export type Claim = {
  claimBenefitAmt: number;
  claimCOBAmt: number;
  claimCoInsAmt: number;
  claimCopayAmt: number;
  claimDeductibleAmt: number;
  claimHRAIndicator: string;
  claimHighServiceDate: number;
  claimId: string;
  relatedClaimId: string;
  claimIdAdjustedFrom: string;
  claimIdAdjustedTo: string;
  claimIdCreatedFrom: string;
  claimLowServiceDate: number;
  claimNonCoveredAmt: number;
  networkSavings: number;
  claimPaidAmt: number;
  claimPaidDate: number;
  claimPatientOweAmt: number;
  claimRecievedDate: number;
  claimStatusCode: string;
  claimStatusDescription: string;
  claimSubType: string;
  claimTotalChargeAmt: number;
  claimType: string;
  hraClaimPaidAmount: number;
  bcbstPaysYou: number;
  memberCk: number;
  providerId: string;
  providerName: string;
  providerSpecialty: string;
  cryptedKeyForEOB: string;
  networkID: string;
  networkSummary: string;
  tierDescription: string;
  tierId: string;
  tierPrefix: string;
  accountNumber: number;
  totalAllowedAmt: number;
  providerBillingTIN: string;
  payToProviderAddress1: string;
  payToProviderCity: string;
  payToProviderState: string;
  payToProviderZip: string;
  patientAccNo: string;
  vndAmtPaid: number;
  providerType: string;
  eobInd: string;
  memberSuffix: number;
  claimHighServiceCalendarDate: string;
  claimLowServiceCalendarDate: string;
  claimPaidCalendarDate: string;
  claimReceivedCalendarDate: string;
  bcbstpaysYou: number;
};

export type ClaimDetailResponse = {
  claim: Claim;
};